Create gmail.py
This script is written for use inside python scripts so I can be notified by email of progress, completion, or failures. Gmail limits the number of emails you can send in a given day, be aware of that and don't get yourself in trouble with this.  It allows me to set complex jobs to run without the need to babysit them. Probably the most important evolution in my business process. It is called repeatedly by other scripts.