Create Table_to_PointsFeature.py
Use to pull data from a table class, put it on a map, and export it to a feature class. Short hand hack for use inside of other scripts. FeatureClassToFeatureClass_conversion was used originally but it took way longer than it should have. That function was replaced with FeatureToPoint_management. 3+hours vs 15 minutes to do the same job.