Create GIS_Python_Toolbox.pyt
This toolbox is written for Mississippi. 
This tool allows the user to select counties from a\
list that the map will select, zoom to, and mask \
the rest of the state. Two layers are required for\
this script to run. One is named 'County' and the\
Other is Named 'Mask'. The County layer should have\
a hollow symbology and the mask should be solid \
with a display transparency. 

Warning: If you are using ArcGIS 10.1 this tool \
behaves in an erratic manner.